done with 4.15-4.23

+++ Osa4/blogilista/controllers/blogs.js
@@ -1,21 +1,32 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+  .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
   })
   
 blogsRouter.post('/', async (request, response) => {
+  if (!request.decodedToken) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
   try {
-    const { title, author, url, likes } = request.body
+    const { title, author, url, likes} = request.body
+    
+    const user = request.user
     const blog = new Blog({
       title,
       author,
       url,
-      likes: likes || 0
+      likes: likes || 0,
+      user: user._id
     })
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.status(201).json(savedBlog)
   } catch (error) {
     response.status(400).end()
@@ -23,6 +34,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  if (!request.decodedToken) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+  if(!blog) {
+    return response.status(204).end()
+  }
+
+  if (blog.user.toString() !== request.user._id.toString()) {
+    return response.status(401).json({ error: 'wrong user' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
